completed delete function in profile page

+++ Services/items.js
@@ -42,9 +42,14 @@ export const updateItem = async (itemId, itemData) => {
   }
 };
 
-export const deleteItem = async (itemId) => {
+export const deleteItem = async (userId,itemId) => {
   try {
-    const response = await api.delete(`/items/${itemId}`);
+    const token = localStorage.getItem("token");
+    const response = await api.delete(`/items/${userId}/${itemId}`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
     return response.data;
   } catch (error) {
     console.error(error);
